docs: outline frontend refactor and share network color
