+++ packages/nextjs/hooks/scaffold-eth/useNetworkColor.ts
@@ -1,23 +1,11 @@
-import { useTargetNetwork } from "./useTargetNetwork";
-import { useTheme } from "next-themes";
 import { useSelectedNetwork } from "~~/hooks/scaffold-eth";
-import { AllowedChainIds, ChainWithAttributes } from "~~/utils/scaffold-eth";
-
-export const DEFAULT_NETWORK_COLOR: [string, string] = ["#666666", "#bbbbbb"];
-
-export function getNetworkColor(network: ChainWithAttributes, isDarkMode: boolean) {
-  const colorConfig = network.color ?? DEFAULT_NETWORK_COLOR;
-  return Array.isArray(colorConfig) ? (isDarkMode ? colorConfig[1] : colorConfig[0]) : colorConfig;
-}
+import { AllowedChainIds } from "~~/utils/scaffold-eth";
+import { useNetworkColor as useSharedNetworkColor } from "../common/useNetworkColor";
 
 /**
- * Gets the color of the target network
+ * Gets the color of the selected EVM network
  */
 export const useNetworkColor = (chainId?: AllowedChainIds) => {
-  const { resolvedTheme } = useTheme();
-
   const chain = useSelectedNetwork(chainId);
-  const isDarkMode = resolvedTheme === "dark";
-
-  return getNetworkColor(chain, isDarkMode);
+  return useSharedNetworkColor(chain);
 };
