+++ packages/nextjs/hooks/scaffold-stark/useNetworkColor.ts
@@ -1,29 +1,10 @@
 import { useTargetNetwork } from "./useTargetNetwork";
-import { useTheme } from "next-themes";
-import { ChainWithAttributes } from "~~/utils/scaffold-stark";
-
-export const DEFAULT_NETWORK_COLOR: [string, string] = ["#666666", "#bbbbbb"];
-
-export function getNetworkColor(
-  network: ChainWithAttributes,
-  isDarkMode: boolean,
-) {
-  const colorConfig = network.color ?? DEFAULT_NETWORK_COLOR;
-  return Array.isArray(colorConfig)
-    ? isDarkMode
-      ? colorConfig[1]
-      : colorConfig[0]
-    : colorConfig;
-}
+import { useNetworkColor as useSharedNetworkColor } from "../common/useNetworkColor";
 
 /**
- * Gets the color of the target network
+ * Gets the color of the target Starknet network
  */
 export const useNetworkColor = () => {
-  const { resolvedTheme } = useTheme();
   const { targetNetwork } = useTargetNetwork();
-
-  const isDarkMode = resolvedTheme === "dark";
-
-  return getNetworkColor(targetNetwork, isDarkMode);
+  return useSharedNetworkColor(targetNetwork);
 };
